requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
comet-ml==3.3.2
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dulwich==0.20.18
entrypoints==0.3
everett==1.0.3
h5py==3.2.1
idna==2.10
importlib-metadata==3.4.0
ipdb==0.13.4
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.9
notebook==6.2.0
numpy==1.20.1
nvidia-ml-py3==7.352.0
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.15
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pynvml==8.0.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-ignite==0.4.3
pytz==2021.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-toolbelt==0.9.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
synbols==1.0.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.1
zipp==3.4.0
qj==0.2.0
